warning 해결

+++ GTLEngine/Source/Core/DirectX11/DirectXHandle.cpp
@@ -743,8 +743,8 @@ void UDirectXHandle::UpdateWorldProjectionMatrix(ACamera* Camera)
 }
 
 void UDirectXHandle::ResizeViewport(int width, int height) {
-	ViewportInfo.Width = width;
-	ViewportInfo.Height = height;
+	ViewportInfo.Width = static_cast<float>(width);
+	ViewportInfo.Height = static_cast<float>(height);
 }
 
 HRESULT UDirectXHandle::ResizeWindow(int width, int height) {
